edited table

+++ src/components/TasksTableTeacher.js
@@ -16,7 +16,7 @@ import { useEffect } from 'react'
 import { LinearProgress } from '@mui/material';
 import { format, fromUnixTime } from 'date-fns';
 import {deleteTask} from '../utils/mutations'
-import DeleteIcon from '@mui/icons-material/Delete';
+import MoreVertIcon from '@mui/icons-material/MoreVert';
 import IconButton from '@mui/material/IconButton';
 import PropTypes from 'prop-types';
 import Box from '@mui/material/Box';
@@ -35,29 +35,20 @@ function percentDone(task) {
    return ((numConfirmed) / (numCompleted + numConfirmed + numAssigned)) * 100;
 }
 
-function LinearProgressWithLabel(props) {
+function LinearProgressWithLabel({task}) {
+   console.log(task);
    return (
      <Box sx={{ display: 'flex', alignItems: 'center' }}>
-       <Box sx={{ width: '100%', mr: 1 }}>
-         <LinearProgress variant="determinate" {...props} />
+      <Box sx={{ minWidth: 100 }}>
+         <Typography variant="body2" color="text.secondary">{`${task.confirmed?.length}/${task.completed?.length + task.assigned?.length + task.confirmed?.length} students`}</Typography>
        </Box>
-       <Box sx={{ minWidth: 35 }}>
-         <Typography variant="body2" color="text.secondary">{`${Math.round(
-           props.value,
-         )}%`}</Typography>
+       <Box sx={{ minWidth: '50%', mr: 1 }} align="right">
+         <LinearProgress variant="determinate" value={percentDone(task)} />
        </Box>
      </Box>
    );
- }
- 
- LinearProgressWithLabel.propTypes = {
-   /**
-    * The value of the progress indicator for the determinate and buffer variants.
-    * Value between 0 and 100.
-    */
-   value: PropTypes.number.isRequired,
  };
- 
+
 
 export default function TasksTableTeacher({ classroom }) {
 
@@ -79,8 +70,12 @@ export default function TasksTableTeacher({ classroom }) {
       console.log(tasks);
    }, []);
 
+
    const handleDelete = (task) => {
-      deleteTask(classroom.id, task.id).catch(console.error);
+      // message box to confirm deletion
+      if (window.confirm("Are you sure you want to delete this task?")) {
+         deleteTask(classroom.id, task.id).catch(console.error);
+      }
    }
 
    return (
@@ -89,7 +84,7 @@ export default function TasksTableTeacher({ classroom }) {
             <Table sx={{ minWidth: 650 }} aria-label="simple table">
                <TableHead>
                   <TableRow>
-                     <TableCell></TableCell>
+                     <TableCell sx={{m:'1%', p:'1%'}}></TableCell>
                      <TableCell>Task</TableCell>
                      <TableCell>Description</TableCell>
                      <TableCell>Deadline</TableCell>
@@ -101,10 +96,10 @@ export default function TasksTableTeacher({ classroom }) {
                   {tasks?.map((task) => (
                      <TableRow
                         key={task.id}
-                        sx={{ '&:last-child td, &:last-child th': { border: 0 } }}
+                        sx={{ '&:last-child td, &:last-child th': { border: 0 }}}
                      >
 
-                        <TableCell sx={{ "paddingTop": 0, "paddingBottom": 0 }} align="left">
+                        <TableCell sx={{ "paddingTop": 0, "paddingBottom": 0, width: .01 }} align="left">
                            <TaskModalTeacher task={task} classroom={classroom} />
                         </TableCell>
 
@@ -112,8 +107,12 @@ export default function TasksTableTeacher({ classroom }) {
                         <TableCell align="left">{truncate(task.description)}</TableCell>
                         <TableCell align="left">{format(fromUnixTime(task.due), 'MM/dd/yyyy')}</TableCell>
                         <TableCell alight="left">{task.reward}</TableCell>
-                        <TableCell align="left"><LinearProgressWithLabel variant="determinate" value={percentDone(task)} /></TableCell>
-                        <TableCell align="left"><IconButton onClick={() => handleDelete(task)}><DeleteIcon /></IconButton></TableCell>
+                        <TableCell align="left"><LinearProgressWithLabel variant="determinate" task={task} /></TableCell>
+
+                        <TableCell align="right" sx={{width:.01}}>
+                           <IconButton onClick={() => handleDelete(task)}><MoreVertIcon /></IconButton>
+                        </TableCell>
+
                      </TableRow>
                   ))}
                </TableBody>
